refacto(backend): regrouped config related handling
