+++ scripts/config/config-schema.json
