+++ scripts/server-backend.js
@@ -1,6 +1,6 @@
 const path = require("path");
 
-const config = require("./config");
+const config = require("./config/config");
 const WhiteboardServerSideInfo = require("./WhiteboardServerSideInfo");
 
 function startBackendServer(port) {
