+++ scripts/utils.js
@@ -1,12 +1,3 @@
-const path = require("path");
-const fs = require("fs");
-const yaml = require("js-yaml");
-
-const Ajv = require("ajv");
-const ajv = new Ajv({ allErrors: true });
-
-const configSchema = require("./config-schema.json");
-
 function getArgs() {
     const args = {};
     process.argv.slice(2, process.argv.length).forEach((arg) => {
@@ -26,85 +17,4 @@ function getArgs() {
     return args;
 }
 
-/**
- * TODO
- *
- * @param path
- * @return {any}
- */
-function getConfig(path) {
-    return yaml.safeLoad(fs.readFileSync(path, "utf8"));
-}
-
-/**
- * TODO
- * @param config
- * @param warn
- * @return {*}
- */
-function isConfigValid(config, warn = true) {
-    const validate = ajv.compile(configSchema);
-    const isValidAgainstSchema = validate(config);
-
-    if (!isValidAgainstSchema && warn) console.warn(validate.errors);
-
-    let structureIsValid = false;
-    try {
-        structureIsValid = config.frontend.performance.pointerEventsThrottling.some(
-            (item) => item.fromNbUser === 0
-        );
-    } catch (e) {
-        if (!e instanceof TypeError) {
-            throw e;
-        }
-    }
-
-    if (!structureIsValid && warn)
-        console.warn(
-            "At least one item under frontend.performance.pointerEventsThrottling" +
-                "must have fromNbUser set to 0"
-        );
-
-    return isValidAgainstSchema && structureIsValid;
-}
-
-/**
- * TODO
- * @return {*}
- */
-function getDefaultConfig() {
-    const defaultConfigPath = path.join(__dirname, "..", "config.default.yml");
-    return getConfig(defaultConfigPath);
-}
-
-/**
- * TODO
- * Deep merges objects, not arrays.
- *
- * @param baseConfig
- * @param overrideConfig
- * @return {{}}
- */
-function deepMergeConfigs(baseConfig, overrideConfig) {
-    const out = {};
-
-    Object.entries(baseConfig).forEach(([key, val]) => {
-        out[key] = val;
-        if (overrideConfig.hasOwnProperty(key)) {
-            const overrideVal = overrideConfig[key];
-            if (typeof val === "object" && !Array.isArray(val) && val !== null) {
-                out[key] = deepMergeConfigs(val, overrideVal);
-            } else {
-                out[key] = overrideVal;
-            }
-        }
-    });
-
-    return out;
-}
-
 module.exports.getArgs = getArgs;
-module.exports.getConfig = getConfig;
-module.exports.getDefaultConfig = getDefaultConfig;
-module.exports.deepMergeConfigs = deepMergeConfigs;
-module.exports.isConfigValid = isConfigValid;
